refactor browser archive extraction
